Respect aliases during dependency discovery
So that the backbone:exoskeleton alias means backbone isn't included in
the final bundle.
