+++ lib/deppack/explore.js
@@ -1,10 +1,8 @@
 'use strict';
 
 const sysPath = require('path');
-const glob = require('glob');
 const each = require('async-each');
 const detective = require('detective');
-const promisify = require('../helpers').promisify;
 const mediator = require('../mediator');
 const shims = require('./shims');
 const helpers = require('./helpers');
@@ -27,35 +25,6 @@ const fileModsNotChanged = (file, mods) => {
   return curr.join(',') === mods.join(',');
 };
 
-let modMap;
-
-const buildModMap = () => {
-  if (modMap) return Promise.resolve();
-
-  // this is needed to not throw on app-wise requires
-  // note that fileList.files can't be used because it's not fully populated until the first compilation
-  const topPaths = mediator.paths.watched.join(',');
-  return promisify(glob)(`{${topPaths}}/**/*`).then(mods => {
-    const modMap0 = mods.reduce((map, mod) => {
-      const name = mediator.nameCleaner(mod.split('.').slice(0, -1).join('.'));
-      map[name] = mod;
-      return map;
-    }, {});
-
-    const shimModMap = shims.emptyShims.filter(x => mediator.packages[x] == null).reduce((acc, shim) => {
-      acc[shim] = shims.makeSpecialShimFname(shim);
-      return acc;
-    }, {});
-
-    const shimFileMap = Object.keys(shims.fileShims).filter(x => mediator.packages[x] == null).reduce((acc, shim) => {
-      acc[shim] = shims.fileShims[shim];
-      return acc;
-    }, {});
-
-    modMap = Object.assign({}, modMap0, shimModMap, shimFileMap);
-  });
-};
-
 const isJs = (fileList, path) => fileList.find(path).type === 'javascript';
 
 const packageRe = /node_modules/;
@@ -75,7 +44,7 @@ const exploreDeps = fileList => {
     const deps = isApp(path) ? allDeps.filter(x => !helpers.isRelative(x)) : allDeps;
 
     const resPath = sysPath.resolve(path);
-    const res = (i, cb) => xBrowserResolve(i, {filename: resPath, modules: modMap}, cb);
+    const res = (i, cb) => xBrowserResolve(i, {filename: resPath}, cb);
 
     return new Promise((resolve, reject) => {
       if (fileModsNotChanged(path, deps)) return resolve(x);
@@ -113,6 +82,5 @@ const exploreDeps = fileList => {
 };
 
 module.exports = exploreDeps;
-module.exports.buildModMap = buildModMap;
 module.exports.addFileMods = addFileMods;
 module.exports.addFileFiles = addFileFiles;
