+++ lib/deppack/init.js
@@ -14,6 +14,7 @@ const loadInit = (config, json) => new Promise((resolve, reject) => {
   mediator.packages = json.dependencies || {};
   mediator.nameCleaner = config.modules.nameCleaner;
   mediator.paths = config.paths;
+  mediator.npm = config.npm || {};
 
   const rootPath = sysPath.resolve(mediator.paths.root);
 
@@ -22,7 +23,7 @@ const loadInit = (config, json) => new Promise((resolve, reject) => {
 
   const res = (i, cb) => xBrowserResolve(i, {filename: globalPseudofile}, cb);
 
-  explore.buildModMap().then(() => {
+  xBrowserResolve.buildModMap().then(() => {
     const globModules = Object.keys(globals).map(k => globals[k]);
     each(globModules, res, (err, fullPaths) => {
       if (err) return reject(err);
