+++ lib/deppack/resolve.js
@@ -1,14 +1,47 @@
 'use strict';
 
 const browserResolve = require('browser-resolve');
+const glob = require('glob');
 const modules = require('./modules');
 const globalPseudofile = require('./helpers').globalPseudofile;
 const mediator = require('../mediator');
+const promisify = require('../helpers').promisify;
+const shims = require('./shims');
+
+let modMap;
+
+const buildModMap = () => {
+  if (modMap) return Promise.resolve();
+
+  // this is needed to not throw on app-wise requires
+  // note that fileList.files can't be used because it's not fully populated until the first compilation
+  const topPaths = mediator.paths.watched.join(',');
+  return promisify(glob)(`{${topPaths}}/**/*`).then(mods => {
+    const modMap0 = mods.reduce((map, mod) => {
+      const name = mediator.nameCleaner(mod.split('.').slice(0, -1).join('.'));
+      map[name] = mod;
+      return map;
+    }, {});
+
+    const shimModMap = shims.emptyShims.filter(x => mediator.packages[x] == null).reduce((acc, shim) => {
+      acc[shim] = shims.makeSpecialShimFname(shim);
+      return acc;
+    }, {});
+
+    const shimFileMap = Object.keys(shims.fileShims).filter(x => mediator.packages[x] == null).reduce((acc, shim) => {
+      acc[shim] = shims.fileShims[shim];
+      return acc;
+    }, {});
+
+    modMap = Object.assign({}, modMap0, shimModMap, shimFileMap);
+  });
+};
 
 const resolveErrorRe = /Cannot find module '(.+)' from '(.+)'/;
 
-module.exports = (mod, opts, cb) => {
-  Object.assign(opts, {packageFilter: modules.applyPackageOverrides});
+const resolve = (mod, opts, cb) => {
+  Object.assign(opts, {packageFilter: modules.applyPackageOverrides, modules: modMap});
+  mod = mediator.npm.aliases && mediator.npm.aliases[mod] || mod;
   browserResolve(mod, opts, (err, res) => {
     if (err) {
       if (resolveErrorRe.test(err.message)) {
@@ -36,3 +69,6 @@ module.exports = (mod, opts, cb) => {
     cb(null, res);
   });
 };
+
+module.exports = resolve;
+module.exports.buildModMap = buildModMap;
