+++ lib/mediator.js
@@ -4,3 +4,4 @@ exports.packages = {};
 exports.conventions = {};
 exports.paths = [];
 exports.nameCleaner = x => x;
+exports.npm = {};
